Create/Move Module to Admin Folder #1 - move all module to (admin) folder except login - add tooltip for button actions in datatable
