+++ app/(admin)/manages/products/page.tsx
@@ -1,0 +1,23 @@
+//- app/(admin)/manages/products/page.tsx
+
+import { AppHeader } from "@/components/core/app-layout"
+import { MaintenancePage } from "@/components/core/maintenance"
+
+const breadcrumbItems = [
+  {
+    label: "Manages",
+  },
+  {
+    label: "Products",
+  },
+]
+
+export default function ManagesProductsPage() {
+  return (
+    <>
+      <AppHeader breadcrumbItems={breadcrumbItems} />
+
+      <MaintenancePage />
+    </>
+  )
+}
