+++ app/(admin)/manages/recipes/page.tsx
@@ -1,0 +1,23 @@
+//- app/(admin)/manages/recipes/page.tsx
+
+import { AppHeader } from "@/components/core/app-layout"
+import { MaintenancePage } from "@/components/core/maintenance"
+
+const breadcrumbItems = [
+  {
+    label: "Manages",
+  },
+  {
+    label: "Recipes",
+  },
+]
+
+export default function ManagesRecipesPage() {
+  return (
+    <>
+      <AppHeader breadcrumbItems={breadcrumbItems} />
+
+      <MaintenancePage />
+    </>
+  )
+}
