+++ app/(admin)/manages/users/page.tsx
@@ -1,0 +1,23 @@
+//- app/(admin)/manages/users/page.tsx
+
+import { AppHeader } from "@/components/core/app-layout"
+import { MaintenancePage } from "@/components/core/maintenance"
+
+const breadcrumbItems = [
+  {
+    label: "Manages",
+  },
+  {
+    label: "Users",
+  },
+]
+
+export default function ManagesUsersPage() {
+  return (
+    <>
+      <AppHeader breadcrumbItems={breadcrumbItems} />
+
+      <MaintenancePage />
+    </>
+  )
+}
