+++ app/(admin)/reports/page.tsx
@@ -1,0 +1,20 @@
+//- app/(admin)/reports/page.tsx
+
+import { AppHeader } from "@/components/core/app-layout"
+import { MaintenancePage } from "@/components/core/maintenance"
+
+const breadcrumbItems = [
+  {
+    label: "Reports",
+  },
+]
+
+export default function ReportsPage() {
+  return (
+    <>
+      <AppHeader breadcrumbItems={breadcrumbItems} />
+
+      <MaintenancePage />
+    </>
+  )
+}
