+++ app/(admin)/sales/invoices/page.tsx
@@ -1,0 +1,23 @@
+//- app/(admin)/sales/invoices/page.tsx
+
+import { AppHeader } from "@/components/core/app-layout"
+import { MaintenancePage } from "@/components/core/maintenance"
+
+const breadcrumbItems = [
+  {
+    label: "Sales",
+  },
+  {
+    label: "Invoices",
+  },
+]
+
+export default function InvoicesPage() {
+  return (
+    <>
+      <AppHeader breadcrumbItems={breadcrumbItems} />
+
+      <MaintenancePage />
+    </>
+  )
+}
