+++ app/(admin)/sales/orders/page.tsx
@@ -1,0 +1,23 @@
+//- app/(admin)/sales/orders/page.tsx
+
+import { AppHeader } from "@/components/core/app-layout"
+import { MaintenancePage } from "@/components/core/maintenance"
+
+const breadcrumbItems = [
+  {
+    label: "Sales",
+  },
+  {
+    label: "Orders",
+  },
+]
+
+export default function OrdersPage() {
+  return (
+    <>
+      <AppHeader breadcrumbItems={breadcrumbItems} />
+
+      <MaintenancePage />
+    </>
+  )
+}
