+++ app/(admin)/sales/shipments/page.tsx
@@ -1,0 +1,23 @@
+//- app/(admin)/sales/shipments/page.tsx
+
+import { AppHeader } from "@/components/core/app-layout"
+import { MaintenancePage } from "@/components/core/maintenance"
+
+const breadcrumbItems = [
+  {
+    label: "Sales",
+  },
+  {
+    label: "Shipments",
+  },
+]
+
+export default function ShipmentsPage() {
+  return (
+    <>
+      <AppHeader breadcrumbItems={breadcrumbItems} />
+
+      <MaintenancePage />
+    </>
+  )
+}
