+++ app/(admin)/settings/page.tsx
@@ -1,0 +1,20 @@
+//- app/(admin)/settings/page.tsx
+
+import { AppHeader } from "@/components/core/app-layout"
+import { MaintenancePage } from "@/components/core/maintenance"
+
+const breadcrumbItems = [
+  {
+    label: "Settings",
+  },
+]
+
+export default function SettingsPage() {
+  return (
+    <>
+      <AppHeader breadcrumbItems={breadcrumbItems} />
+
+      <MaintenancePage />
+    </>
+  )
+}
