--- app/customers/all-customers/page.tsx
@@ -1,23 +1,0 @@
-//- app/customers/all-customers/page.tsx
-
-import { AppHeader } from "@/components/core/app-layout"
-import { MaintenancePage } from "@/components/core/maintenance"
-
-const breadcrumbItems = [
-  {
-    label: "Customers",
-  },
-  {
-    label: "All Customers",
-  },
-]
-
-export default function AllCustomersPage() {
-  return (
-    <>
-      <AppHeader breadcrumbItems={breadcrumbItems} />
-
-      <MaintenancePage />
-    </>
-  )
-}
