--- app/customers/groups/page.tsx
@@ -1,23 +1,0 @@
-//- app/customers/groups/page.tsx
-
-import { AppHeader } from "@/components/core/app-layout"
-import { MaintenancePage } from "@/components/core/maintenance"
-
-const breadcrumbItems = [
-  {
-    label: "Customers",
-  },
-  {
-    label: "Groups",
-  },
-]
-
-export default function GroupsPage() {
-  return (
-    <>
-      <AppHeader breadcrumbItems={breadcrumbItems} />
-
-      <MaintenancePage />
-    </>
-  )
-}
