--- app/customers/segments/page.tsx
@@ -1,23 +1,0 @@
-//- app/customers/segments/page.tsx
-
-import { AppHeader } from "@/components/core/app-layout"
-import { MaintenancePage } from "@/components/core/maintenance"
-
-const breadcrumbItems = [
-  {
-    label: "Customers",
-  },
-  {
-    label: "Segments",
-  },
-]
-
-export default function SegmentsPage() {
-  return (
-    <>
-      <AppHeader breadcrumbItems={breadcrumbItems} />
-
-      <MaintenancePage />
-    </>
-  )
-}
