--- app/helps/page.tsx
@@ -1,20 +1,0 @@
-//- app/helps/page.tsx
-
-import { AppHeader } from "@/components/core/app-layout"
-import { MaintenancePage } from "@/components/core/maintenance"
-
-const breadcrumbItems = [
-  {
-    label: "Helps",
-  },
-]
-
-export default function HelpsPage() {
-  return (
-    <>
-      <AppHeader breadcrumbItems={breadcrumbItems} />
-
-      <MaintenancePage />
-    </>
-  )
-}
