--- app/manages/products/page.tsx
@@ -1,23 +1,0 @@
-//- app/manages/products/page.tsx
-
-import { AppHeader } from "@/components/core/app-layout"
-import { MaintenancePage } from "@/components/core/maintenance"
-
-const breadcrumbItems = [
-  {
-    label: "Manages",
-  },
-  {
-    label: "Products",
-  },
-]
-
-export default function ManagesProductsPage() {
-  return (
-    <>
-      <AppHeader breadcrumbItems={breadcrumbItems} />
-
-      <MaintenancePage />
-    </>
-  )
-}
