--- app/socials/quotes/column.tsx
@@ -1,37 +1,0 @@
-//- app/socials/quotes/column.ts
-
-import { ColumnDef } from "@tanstack/react-table"
-import { SocialQuoteList } from "./type"
-import {
-  actionColHeader,
-  actionColKey,
-  actionColSize,
-  CellActions,
-} from "@/components/core/data-table/columns"
-
-export const Columns: ColumnDef<SocialQuoteList>[] = [
-  {
-    accessorKey: "quote",
-    header: "Quote",
-    size: 0,
-  },
-  {
-    accessorKey: "author",
-    header: "Author",
-    size: 250,
-  },
-  {
-    accessorKey: actionColKey,
-    header: actionColHeader,
-    size: actionColSize,
-    enableSorting: false,
-    cell: ({ row }) => {
-      return (
-        <CellActions
-          pathEdit={`/socials/quotes/edit/${row.original.id}`}
-          pathDelete={`/socials/quotes/delete/${row.original.id}`}
-        />
-      )
-    },
-  },
-]
